警告の修正。

+++ DirectX 初ゲーム/Lerp.cpp
@@ -31,7 +31,7 @@ Color LerpC(Color a, Color b, float t)
 		Color result = 0;
 		for (int i = 0; i < 4; i++)
 		{
-			result += (int)Lerpf((a & (0x000000ff << (8 * i))) >> (8 * i), (b & (0x000000ff << (8 * i))) >> (8 * i), t) << (8 * i);
+			result += (int)Lerpf((float)((a & (0x000000ff << (8 * i))) >> (8 * i)), (float)((b & (0x000000ff << (8 * i))) >> (8 * i)), t) << (8 * i);
 		}
 		return result;
 	}
